backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
cryptography==42.0.5
python-dotenv
requests
httpx
shopifyapi
qdrant-client
langchain
langgraph
google-generativeai
langchain-google-genai
langchain-community
pydantic[email]
email-validator
asyncio
stream-chat
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
authlib
itsdangerous
pyotp
qrcode[pil]
redis
celery
flower
